moved `HTMLHead` to `hooks/meta`
