+++ app/src/layout/JustSCOTUS/JustSCOTUS.tsx
@@ -8,7 +8,7 @@ import SearchResult from 'src/components/ui/SearchResult';
 import Button from 'src/components/ui/Button';
 import DropdownMenu, { ValueLabelPair } from 'src/components/ui/DropdownMenu';
 import Sidebar from 'src/components/ui/Sidebar';
-import { useTitle } from 'src/HTMLHead';
+import { useTitle } from 'src/hooks/meta';
 import { SCOTUSKit } from 'src/utils/JustSDK';
 import './JustSCOTUS.css';
