+++ app/src/layout/JustSearch/SearchEngine.tsx
@@ -6,7 +6,7 @@ import LoadingAnimation from 'src/components/LoadingAnimation';
 import SearchResult from 'src/components/ui/SearchResult';
 import SearchBar from 'src/components/ui/SearchBar';
 import Sidebar from 'src/components/ui/Sidebar';
-import { useTitle } from 'src/HTMLHead';
+import { useTitle } from 'src/hooks/meta';
 import { SERPKit } from 'src/utils/JustSDK';
 import './SearchEngine.css';
